refactor(listings): use shared queries for server pages and api
