+++ app/listings/[id]/page.tsx
@@ -1,57 +1,27 @@
-import { headers } from "next/headers";
 import { notFound } from "next/navigation";
 
 import { getOptionalSession } from "@/lib/auth/session";
+import { readListingById, toListingReadContext } from "@/lib/listings/queries";
 import { ListingDetails } from "@/components/listing-details/ListingDetails";
 import type { ListingDetailProps } from "@/components/listing-details/ListingDetails";
-import {
-  listingByIdResponseSchema,
-  type ListingDetails as ListingDetailsData,
-} from "@/shared/schemas/listings";
+import type { ListingDetails as ListingDetailsData } from "@/shared/schemas/listings";
 
 type PageProps = {
   params: Promise<{ id: string }>;
 };
 
-async function getListingFromApi(id: string): Promise<ListingDetailsData> {
-  const requestHeaders = await headers();
-  const host = requestHeaders.get("x-forwarded-host") ?? requestHeaders.get("host");
-  const protocol = requestHeaders.get("x-forwarded-proto") ?? "http";
-  const cookieHeader = requestHeaders.get("cookie");
-  const allowedHosts = new Set([
-    "localhost:3000",
-    "localhost:3100",
-    "localhost:3200",
-    "localhost:3201",
-  ]);
-  const trustedBaseUrl = process.env.NEXTAUTH_URL;
+async function getListingDetails(id: string): Promise<ListingDetailsData> {
   const optionalSession = await getOptionalSession();
-  const shouldForwardPreviewSession =
-    optionalSession.authzUser?.role === "admin" || optionalSession.authzUser?.role === "partner";
-
-  const baseUrl =
-    trustedBaseUrl || (host && allowedHosts.has(host) ? `${protocol}://${host}` : null);
-
-  if (!baseUrl) {
-    notFound();
-  }
-
-  const response = await fetch(`${baseUrl}/api/listings/${id}`, {
-    cache: "no-store",
-    // Writers can preview non-public listings through the same details page.
-    headers: shouldForwardPreviewSession && cookieHeader ? { cookie: cookieHeader } : undefined,
+  const details = await readListingById({
+    id,
+    auth: toListingReadContext(optionalSession),
   });
 
-  if (response.status === 404) {
+  if (!details) {
     notFound();
   }
 
-  if (!response.ok) {
-    throw new Error("Failed to load listing details");
-  }
-
-  const payload = listingByIdResponseSchema.parse(await response.json());
-  return payload.data;
+  return details;
 }
 
 function mapListingDetailsToDisplay(details: ListingDetailsData): ListingDetailProps {
@@ -77,7 +47,7 @@ function mapListingDetailsToDisplay(details: ListingDetailsData): ListingDetailP
 
 export default async function ListingDetailsPage({ params }: Readonly<PageProps>) {
   const { id } = await params;
-  const details = await getListingFromApi(id);
+  const details = await getListingDetails(id);
   const displayDetails = mapListingDetailsToDisplay(details);
 
   return <ListingDetails {...displayDetails} />;
